requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
passlib[bcrypt]
python-jose